修复BUG，第三方全部采用cocoapods集成

+++ WYBasisKit/WYBasisKit/Practical/NSString/NSString+Extension.m
@@ -138,21 +138,21 @@
      // 秒转小时
      NSInteger hours = time/3600;
      if (hours<24) {
-     return [NSString stringWithFormat:@"%ld小时前",hours];
+         return [NSString stringWithFormat:@"%ld小时前",(long)hours];
      }
      //秒转天数
      NSInteger days = time/3600/24;
      if (days < 30) {
-     return [NSString stringWithFormat:@"%ld天前",days];
+         return [NSString stringWithFormat:@"%ld天前",(long)days];
      }
      //秒转月
      NSInteger months = time/3600/24/30;
      if (months < 12) {
-     return [NSString stringWithFormat:@"%ld月前",months];
+         return [NSString stringWithFormat:@"%ld月前",(long)months];
      }
      //秒转年
      NSInteger years = time/3600/24/30/12;
-     return [NSString stringWithFormat:@"%ld年前",years];
+    return [NSString stringWithFormat:@"%ld年前",(long)years];
     
      return [self convertToDate:timerSp];
 }
@@ -230,7 +230,7 @@
     if ((nowMonth > month.intValue) || (nowMonth == month.intValue && nowDay >= day.intValue)) {
         userAge++;
     }
-    return [NSString stringWithFormat:@"%ld",userAge];
+    return [NSString stringWithFormat:@"%ld",(long)userAge];
 }
 
 /** 获取手机时间戳 */
